backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
spacy==3.7.2
nltk==3.8.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
django-jazzmin==3.0.1
whitenoise==6.7.0
gunicorn==21.2.0
